Rename "benchmarks/" -> "workloads/" (#32)
**Issue:**
The word "benchmark" is super overloaded (do you mean the workload? the act of running of the workload with a particular library? the results?).

**Description of changes:**
Use the term "workload" for these json files that describe the work we're benchmarking.

**Non-changes:**
I'm holding back from revamping all the runner code, where we call these "BenchmarkConfig". I tried doing some find/replace but it was taking a long time and didn't necessarily read any better. It's already disambiguated there, where we always call it "config".