CDK orchestrator job (#35)
Got the orchestrator working! It successfully runs all the per-instance jobs one after another, with some nice logging while it waits.

The per-instance jobs now have an argument-parsing skeleton, but the real working guts are still TODO.

Customizable parameters are successfully passed from orchestrator to per-instance job. The CDK script sets defaults parameters for each job.

So far, I've got it so you can run the job scripts locally too, which really helps with iteration.

The hardest part of this for me was figuring out how to set up roles for the jobs, and getting the permissions exactly right.