Add s3-benchrunner-cpp, which shares some code with s3-benchrunner-c
I'm making the, admittedly weird, choice to add this as its own project, rather than living under s3-benchrunner-c/. The main reason is: we work directly with the core s3-benchrunner-c/ project a LOT when testing changes to aws-c-s3, and I don't want any weird build complexity getting in the way.

To share code, the cpp/ project includes 1 shared .cpp and .h file from the other folder, and a bit of copy/paste in the CMakeLists.txt. If this gets to be a pain in the future, we can combine them.