Add benchmark runner for aws-sdk-cpp (#59)
Add s3-benchrunner-cpp, which shares some code with s3-benchrunner-c

I'm making the (admittedly weird) choice to add this as its own project, rather than being one big project under `s3-benchrunner-c/`. The main reason is: we work directly with the core `s3-benchrunner-c` project a LOT when testing changes to aws-c-s3, and I don't want any weird build complexity getting in the way. To share code, the cpp/ project uses 1 shared .cpp and .h file from the other folder, and a bit of copy/paste in the CMakeLists.txt. If this gets to be a pain in the future, we can combine them.

Big thanks to @TingDaoK who got a lot of this working in https://github.com/awslabs/aws-crt-s3-benchmarks/pull/51. This PR is different in that: projects live in separate directories, and use c++ templates to share code between the `Aws::S3Crt::S3CrtClient` and `Aws::S3::S3Client` runners.

Co-authored-by: @TingDaoK